% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zht.R
\name{zht-class}
\alias{[.zht}
\alias{[<-.zht}
\alias{haskey}
\alias{zht-class}
\title{Zobrist Hash Table Class}
\usage{
\method{[}{zht}(x, state, ...) <- value

\method{[}{zht}(x, state, ...)

haskey(x, ...)
}
\arguments{
\item{x}{object}

\item{state}{state object}

\item{...}{additional argument to be passed to state->key conversion}

\item{value}{value to be inserted}
}
\value{
\itemize{
  \item{\code{`[<-`} and \code{`[`} returns reference to the object}
  \item{\code{haskey} returns logical indicating the existing of the state}
}
}
\description{
Insert, get, find, and delete methods for Zobrist hash table
}
\details{
to be added
}
\examples{
z <- zht(5, 4)

z[1:3]
z[1:3] <- 15
haskey(z, 1:3)
z[1:3] <- NULL  # deletion
haskey(z, 1:3)
}

