% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zht.R
\name{zht}
\alias{zht}
\title{Create a Zobrist Hash Table}
\usage{
zht(keysize, convfunc = NULL, hashsize = 10, rehashable = FALSE,
  threslf = 0.9)
}
\arguments{
\item{keysize}{Positive integer. Bit size of keys}

\item{convfunc}{Function that converts state into key.
If not specified, it is assumed that the state is identical to key.
If specified, the function must take \code{state} as the first entry,
and keyword arguments \code{...}.  See details.}

\item{hashsize}{Positive integer. Bit size of hash values}

\item{rehashable}{Logical. \code{TRUE} if
hashsize should be increased dynamically}

\item{threslf}{Numeric. When \code{rehash = TRUE},
rehashing is implemented when the load factor exceeds this value}
}
\value{
zht object
}
\description{
Create a Zobrist Hash Table
}
\details{
to be added
}

