% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zobristht.R
\name{zobristht}
\alias{zobristht}
\title{Create a Zobrist Hash Table}
\usage{
zobristht(keysize, hashsize, rehashable = FALSE, threslf = 0.9,
  memorysize = 10)
}
\arguments{
\item{keysize}{Positive integer. Bit size of keys}

\item{hashsize}{Positive integer. Bit size of hash values}

\item{threslf}{Numeric. When \code{rehash = TRUE},
rehashing is implemented when the load factor exceeds this value}

\item{memorysize}{Nonnegative integer. This specifies how many recent
hash values to be memorized for a quick access}

\item{rehash}{Logical. \code{TRUE} if
hashsize should be increased dynamically}
}
\value{
zobristhash object
}
\description{
Create a Zobrist Hash Table
}

